New Development: no             JIRA Issue: no       Ready to Release: yes      Interface Changes: none What Interface Changed: none          Test Programs: none   Put in Release Notes: no
           Description: /home/dpetry/temp/casa-nightly-sneeze/code/include/display/DisplayEvents/RegionTool.h:66: error: operator new takes type size_t (unsigned int) as first parameter




git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@16989 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/display/implement/DisplayEvents/RegionTool.h
@@ -63,7 +63,7 @@ namespace casa {
 		    ~State( );
 
 		private:
-		    void *operator new( unsigned long ); // { /* refresh( ) with automatic creation/deletion in mind */ }
+		    void *operator new( size_t ); // { /* refresh( ) with automatic creation/deletion in mind */ }
 		    State( const State & ) { }
 		    State( ) { }
